sally-backend/requirements.txt
# Sally TSM Backend - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# Database Drivers
psycopg2-binary==2.9.9      # PostgreSQL
pymysql==1.1.0              # MySQL
cx-Oracle==8.3.0            # Oracle
aiosqlite==0.19.0           # SQLite async
asyncpg==0.29.0             # PostgreSQL async
aiomysql==0.2.0             # MySQL async

# Google Gemini AI
google-generativeai==0.3.1

# Security & Utilities
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4           # Password hashing
python-multipart==0.0.6          # File uploads
aiofiles==23.2.1                 # Async file operations

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
